LLMPersistence/requirements.txt
pika==1.3.2
langchain==0.0.305
chromadb==0.4.12
urllib3==2.0.4
python-dotenv==1.0.0
unstructured==0.10.8
extract-msg==0.45.0
tabulate==0.9.0
pandoc==2.3
pypandoc==1.11
tqdm==4.66.1
sentence_transformers==2.2.2